+++ server/lib/tuist_web/live/tests_live.ex
@@ -211,6 +211,7 @@ defmodule TuistWeb.TestsLive do
           case run.status do
             "success" -> "var:noora-chart-primary"
             "failure" -> "var:noora-chart-destructive"
+            "skipped" -> "var:noora-chart-warning"
           end
 
         value = (run.duration / 1000) |> Decimal.from_float() |> Decimal.round(0)