+++ cli/CHANGELOG.md
@@ -1,13 +1,57 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
-## What's Changed in 4.109.1<!-- RELEASE NOTES START -->
+## What's Changed in 4.111.0<!-- RELEASE NOTES START -->
+
+### ⛰️  Features
+
+* report module cache subhashes by [@fortmarek](https://github.com/fortmarek) in [#8822](https://github.com/tuist/tuist/pull/8822)
+### 🐛 Bug Fixes
+
+* respect explicit cache profile none with target focus by [@hiltonc](https://github.com/hiltonc) in [#8830](https://github.com/tuist/tuist/pull/8830)
+
+**Full Changelog**: https://github.com/tuist/tuist/compare/4.110.3...4.111.0
+
+## What's Changed in 4.110.3<!-- RELEASE NOTES START -->
 
 ### 🐛 Bug Fixes
 
+* handle skipped tests due to a failed build by [@fortmarek](https://github.com/fortmarek) in [#8808](https://github.com/tuist/tuist/pull/8808)
+
+**Full Changelog**: https://github.com/tuist/tuist/compare/4.110.2...4.110.3
+
+## What's Changed in 4.110.2<!-- RELEASE NOTES START -->
+
+### 🐛 Bug Fixes
+
+* false positive for a .uiTests implicit import of .app by [@fortmarek](https://github.com/fortmarek) in [#8811](https://github.com/tuist/tuist/pull/8811)
+
+**Full Changelog**: https://github.com/tuist/tuist/compare/4.110.1...4.110.2
+
+## What's Changed in 4.110.1<!-- RELEASE NOTES START -->
+
+### 🐛 Bug Fixes
+
+* duration for test cases with custom label by [@fortmarek](https://github.com/fortmarek) in [#8800](https://github.com/tuist/tuist/pull/8800)
+
+**Full Changelog**: https://github.com/tuist/tuist/compare/4.110.0...4.110.1
+
+## What's Changed in 4.110.0<!-- RELEASE NOTES START -->
+
+### ⛰️  Features
+
+* deprecate tuist build command by [@pepicrft](https://github.com/pepicrft) in [#8401](https://github.com/tuist/tuist/pull/8401)
+
+**Full Changelog**: https://github.com/tuist/tuist/compare/4.109.2...4.110.0
+
+## What's Changed in 4.109.2<!-- RELEASE NOTES START -->
+
+### 🐛 Bug Fixes
+
+* relegate test result upload error to a warning by [@fortmarek](https://github.com/fortmarek) in [#8790](https://github.com/tuist/tuist/pull/8790)
 * Don't replace targeted external dependencies with cached binary by [@hiltonc](https://github.com/hiltonc) in [#8731](https://github.com/tuist/tuist/pull/8731)
 
-**Full Changelog**: https://github.com/tuist/tuist/compare/4.109.0...4.109.1
+**Full Changelog**: https://github.com/tuist/tuist/compare/4.109.0...4.109.2
 
 ## What's Changed in 4.109.0<!-- RELEASE NOTES START -->